requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
boto3==1.18.0
pdf2image==1.16.0
Pillow==8.3.1
